requirements.txt
asgiref==3.8.1
Django==5.0.7
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==1.7.1
sqlparse==0.5.1
tzdata==2024.1
